requirements.txt
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
camel-converter==3.0.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
hiredis==2.2.3
idna==3.4
kombu==5.2.4
meilisearch==0.28.2
packaging==23.0
prompt-toolkit==3.0.36
psycopg==3.1.10
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.6.0
pytz==2022.6
redis==4.6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.0
